add  mongo plugin actions
