+++ src/Plugins/BotSharp.Plugin.MongoStorage/Collections/PluginDocument.cs
@@ -1,0 +1,6 @@
+namespace BotSharp.Plugin.MongoStorage.Collections;
+
+public class PluginDocument : MongoBase
+{
+    public List<string> EnabledPlugins { get; set; }
+}
